fundoo/requirements.txt
alembic==1.0.11
astroid==2.2.5
atomicwrites==1.3.0
attrs==19.1.0
bcrypt==3.1.7
blinker==1.4
cffi==1.12.3
Click==7.0
decorator==4.4.0
Flask==1.1.1
Flask-Bcrypt==0.7.1
Flask-JWT==0.3.2
Flask-Login==0.4.1
Flask-Mail==0.9.1
Flask-Migrate==2.5.2
Flask-ShortUrl==0.2.0
Flask-SQLAlchemy==2.4.0
Flask-WTF==0.14.2
importlib-metadata==0.19
isort==4.3.21
itsdangerous==1.1.0
Jinja2==2.10.1
lazy-object-proxy==1.4.1
Mako==1.1.0
MarkupSafe==1.1.1
mccabe==0.6.1
more-itertools==7.2.0
packaging==19.1
pluggy==0.12.0
psycopg2-binary==2.8.3
public==2019.4.13
py==1.8.0
pycparser==2.19
PyJWT==1.4.2
pylint==2.3.1
pylint-flask-sqlalchemy==0.1.0
pyparsing==2.4.2
pytest==5.0.1
pytest-flask==0.15.0
python-dateutil==2.8.0
python-editor==1.0.4
self==2019.4.13
short-url==1.2.2
six==1.12.0
SQLAlchemy==1.3.6
typed-ast==1.4.0
wcwidth==0.1.7
Werkzeug==0.15.5
wrapt==1.11.2
WTForms==2.2.1
zipp==0.5.2
